refactor:extract method for abstract class
